// src/services/spaceX/types.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */


export interface paths {
  "/rockets/{rocketId}": {
    get: operations["RocketsOne"];
  };
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
    RocketOneResponse: {
      id: number;
      company: string;
      country: string;
      flickr_images: string[];
      cost_per_launch: number;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export interface operations {

  RocketsOne: {
    parameters: {
      path: {
        /** @example falcon9 */
        rocketId: string;
      };
    };
    responses: {
      /** @description Get a rocket */
      200: {
        content: {
          "application/json": components["schemas"]["RocketOneResponse"];
        };
      };
      /** @description Rocket not found */
      404: {
        content: {
          "application/json": {
            error?: string;
          };
        };
      };
    };
  };
}